moved from tests.map to tests.lcs
